style: 예외처리 추가, 정리 자잘한 메소드 수정

+++ src/main/java/com/sprint/mission/discodeit/repository/file/FileChannelRepository.java
@@ -110,6 +110,10 @@ public class FileChannelRepository implements ChannelRepository {
     public void deleteChannel(UUID channelId) {
         Path path = filePathUtil.getChannelFilePath(channelId);
 
+        if (!Files.exists(path)) {
+            throw new IllegalStateException("삭제할 파일 없음");
+        }
+
         try{
             Files.delete(path);
         } catch (IOException e) {
